[gg/53] Extracted map related frontend code to separate package
